# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/permissions": {
      "get": {
        "operationId": "PermissionController_getPermissions",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "all-permissions": {
                    "summary": "All Permissions",
                    "value": {
                      "data": [
                        {
                          "name": "John Doe",
                          "prefix": "string",
                          "method": [
                            "string"
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Permissions"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
...
    "/api/users/me/permissions": {
      "get": {
        "operationId": "MeController_getPermissions",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "permissions": {
                    "summary": "Permissions",
                    "value": {
                      "data": [
                        {
                          "name": "John Doe",
                          "prefix": "string",
                          "method": [
                            "string"
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/users/me/enable-bio": {
      "post": {
        "operationId": "MeController_enableBio",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnableBioDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/users/me/coupon": {
      "get": {
        "operationId": "MeController_Coupons",
        "parameters": [
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "coupons": {
                    "summary": "Coupons",
                    "value": {
                      "data": {
                        "id": 1,
                        "code": "string",
                        "title": "string"
                      },
                      "total": 1
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "operationId": "MeController_Profile",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "user-profile": {
                    "summary": "User Profile",
                    "value": {
                      "data": {
                        "id": 1,
                        "name": "John Doe",
                        "email": "user@example.com",
                        "phone": "+123456789",
                        "verified": true,
                        "active": true,
                        "image": "https://example.com/avatar.png",
                        "Details": {
                          "wallet": 1,
                          "points": 1,
                          "male": true
                        },
                        "createdAt": "2025-06-26T18:17:10.000Z",
                        "deletedAt": "2025-06-26T18:17:10.000Z",
                        "Role": {
                          "id": 1,
                          "roleKey": "string",
                          "name": "John Doe"
                        },
                        "unReadNotifications": 1,
                        "Permissions": [
                          {
                            "name": "John Doe",
                            "prefix": "string",
                            "method": [
                              "string"
                            ]
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "patch": {
        "operationId": "MeController_updateCurrentUser",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "delete": {
        "operationId": "MeController_deleteCurrentUser",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/users/me/change-password": {
      "patch": {
        "operationId": "MeController_updatePassword",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPasswordDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/users/me/locale": {
      "patch": {
        "operationId": "MeController_updateLocale",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Bookspa",
    "description": "",
    "version": "1.0",
    "contact": {
      "name": "Fahd Hakem",
      "url": "www.linkedin.com/in/fahd-hakem",
      "email": "fhakem75@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3030",
      "description": "Local"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Sandbox"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "ACCESS Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "REFRESH Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "PASSWORD_RESET Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VERIFY Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VISITOR Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EnableBioDTO": {
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string"
          }
        },
        "required": [
          "deviceId"
        ]
      },
      "UpdateUserPasswordDTO": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "Default@123",
            "description": "Password must be: 8+ characters, 1 uppercase, 1 lowercase, 1 number, 1 special character"
          },
          "newPassword": {
            "type": "string",
            "example": "Default@123",
            "description": "Password must be: 8+ characters, 1 uppercase, 1 lowercase, 1 number, 1 special character"
          }
        },
        "required": [
          "password",
          "newPassword"
        ]
      },
      "UpdateUserDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "test@test.com"
          },
          "phone": {
            "type": "string",
            "example": "+966 0509999999"
          },
          "deviceId": {
            "type": "string"
          },
          "fcm": {
            "type": "string"
          }
        }
      },
...
    }
  }
}